+++ CHANGELOG.md
@@ -3,5 +3,4 @@ Thank you!
 Put the changelog BELOW the dashes. ANYTHING ABOVE IS IGNORED
 hi
 -----------------
-- Fixed an issue that prevented the Rebrush enchantment from working properly.
-- Fixed Brush animations in the third person moving very rapidly.
+- Added Built-In Block Models for the Coffin and Lithops.